Setup and Registration feature:
Introducing basic data models, view models, repository, and network stubbed server module.

The RegistrationActivity has an onClickListener, which
verifies the information provided by the user and shows Success or Error toasts.
The RegistrationActivity creates a Registration data object
and passes it to the RegistrationViewModel for verification.
The RegistrationViewModel then passes it to the repository to check with the server.
If the server responds true, which here means a success response code, the repository registers the user by creating a User data object and saving it to the database.
Saving into database yet to be implemented. After registering the user, the repository return true to the viewModel to indicate successful verification. This information is passed on to the activity to display the appropriate toast message on screen.

User input form fields yet to be implemented in the activity.